cv_service/requirements.txt
ultralytics
opencv-python-headless
kafka-python
psycopg2-binary
pandas
numpy
streamlit